Consolidate common utilities between test_compact_*_main.py tests.
