+++ deltacat/tests/compute/test_compact_partition_incremental_main.py
@@ -14,12 +14,11 @@ from deltacat.tests.compute.test_util_common import (
     get_rcf,
     PartitionKey,
     PartitionKeyType,
-    BASE_TEST_SOURCE_NAMESPACE,
-    BASE_TEST_SOURCE_TABLE_NAME,
-    BASE_TEST_SOURCE_TABLE_VERSION,
-    BASE_TEST_DESTINATION_NAMESPACE,
-    BASE_TEST_DESTINATION_TABLE_NAME,
-    BASE_TEST_DESTINATION_TABLE_VERSION,
+)
+from deltacat.tests.compute.test_util_common_main import (
+    _add_deltas_to_partition_main,
+    add_late_deltas_to_partition_main,
+    create_src_w_deltas_destination_plus_destination_main,
 )
 from deltacat.compute.compactor.model.compactor_version import CompactorVersion
 from deltacat.tests.test_utils.utils import read_s3_contents
@@ -127,251 +126,6 @@ def temp_dir(tmp_path):
     return str(tmp_path)
 
 
-@pytest.fixture(scope="function")  
-def catalog(temp_dir):
-    return CatalogProperties(root=temp_dir)
-
-
-@pytest.fixture(scope="function")
-def main_deltacat_storage_kwargs(catalog):
-    return {"catalog": catalog}
-
-
-@pytest.fixture(scope="function")
-def cleanup_database(catalog):
-    """Cleanup test databases after each test function"""
-    yield
-    # Perform cleanup after test
-    import shutil
-    try:
-        shutil.rmtree(catalog.root)
-    except (FileNotFoundError, OSError):
-        pass  # Ignore cleanup errors
-
-
-# Helper functions for main storage implementation
-def _create_table_main(
-    namespace: str,
-    table_name: str,
-    table_version: str,
-    sort_keys: Optional[List[Any]],
-    partition_keys: Optional[List[PartitionKey]], 
-    input_deltas: pa.Table,
-    ds_mock_kwargs: Optional[Dict[str, Any]],
-):
-    """
-    Main storage version of _create_table
-    """
-    # Create schema from the input delta including partition key fields
-    schema = input_deltas.schema
-    
-    # Add partition key fields to schema if they're not already present
-    if partition_keys:
-        for pk in partition_keys:
-            field_name = pk.key_name
-            if field_name not in schema.names:
-                # Add partition key field with appropriate type
-                if pk.key_type == PartitionKeyType.INT:
-                    field_type = pa.int32()
-                elif pk.key_type == PartitionKeyType.STRING:
-                    field_type = pa.string()
-                else:
-                    field_type = pa.string()  # Default to string
-                
-                schema = schema.append(pa.field(field_name, field_type))
-    
-    # Convert test partition keys to storage partition keys
-    storage_partition_keys = []
-    if partition_keys:
-        for pk in partition_keys:
-            storage_pk = StoragePartitionKey.of(
-                key=[pk.key_name],
-                transform=IdentityTransform.of(),
-                name=pk.key_name
-            )
-            storage_partition_keys.append(storage_pk)
-    
-    partition_scheme = None
-    if storage_partition_keys:
-        partition_scheme = PartitionScheme.of(
-            keys=PartitionKeyList.of(storage_partition_keys),
-            scheme_id="default_partition_scheme"
-        )
-
-    # Create namespace first
-    try:
-        metastore.create_namespace(namespace=namespace, **ds_mock_kwargs)
-    except Exception:
-        pass  # Namespace might already exist
-    
-    # Create table version (which creates table and stream automatically)
-    table, table_version_obj, stream = metastore.create_table_version(
-        namespace=namespace,
-        table_name=table_name,
-        table_version=table_version,
-        schema=Schema.of(schema=schema),
-        partition_scheme=partition_scheme,
-        **ds_mock_kwargs,
-    )
-    return namespace, table_name, table_version
-
-
-def create_src_table_main(
-    sort_keys: Optional[List[Any]],
-    partition_keys: Optional[List[PartitionKey]],
-    input_deltas: pa.Table,
-    ds_mock_kwargs: Optional[Dict[str, Any]],
-):
-    """
-    Main storage version of create_src_table
-    """
-    source_namespace: str = BASE_TEST_SOURCE_NAMESPACE
-    source_table_name: str = BASE_TEST_SOURCE_TABLE_NAME
-    source_table_version: str = BASE_TEST_SOURCE_TABLE_VERSION
-    return _create_table_main(
-        source_namespace,
-        source_table_name,
-        source_table_version,
-        sort_keys,
-        partition_keys,
-        input_deltas,
-        ds_mock_kwargs,
-    )
-
-
-def create_destination_table_main(
-    sort_keys: Optional[List[Any]],
-    partition_keys: Optional[List[PartitionKey]],
-    input_deltas: pa.Table,
-    ds_mock_kwargs: Optional[Dict[str, Any]],
-):
-    """
-    Main storage version of create_destination_table
-    """
-    destination_namespace: str = BASE_TEST_DESTINATION_NAMESPACE
-    destination_table_name: str = BASE_TEST_DESTINATION_TABLE_NAME
-    destination_table_version: str = BASE_TEST_DESTINATION_TABLE_VERSION
-    return _create_table_main(
-        destination_namespace,
-        destination_table_name,
-        destination_table_version,
-        sort_keys,
-        partition_keys,
-        input_deltas,
-        ds_mock_kwargs,
-    )
-
-
-def _add_deltas_to_partition_main(
-    deltas_ingredients: List[Tuple[pa.Table, DeltaType, Optional[Dict[str, str]]]],
-    partition: Optional[Partition],
-    ds_mock_kwargs: Optional[Dict[str, Any]],
-) -> Tuple[Optional[Delta], int]:
-    all_deltas_length = 0
-    for (delta_data, delta_type, delete_parameters) in deltas_ingredients:
-        staged_delta: Delta = metastore.stage_delta(
-            delta_data,
-            partition,
-            delta_type,
-            entry_params=delete_parameters,
-            **ds_mock_kwargs,
-        )
-        incremental_delta = metastore.commit_delta(
-            staged_delta,
-            **ds_mock_kwargs,
-        )
-        all_deltas_length += len(delta_data) if delta_data else 0
-    return incremental_delta, all_deltas_length
-
-
-def add_late_deltas_to_partition_main(
-    late_deltas: List[Tuple[pa.Table, DeltaType, Optional[Dict[str, str]]]],
-    source_partition: Optional[Partition],
-    ds_mock_kwargs: Optional[Dict[str, Any]],
-) -> Tuple[Optional[Delta], int]:
-    return _add_deltas_to_partition_main(late_deltas, source_partition, ds_mock_kwargs)
-
-
-def create_src_w_deltas_destination_plus_destination_main(
-    sort_keys: Optional[List[Any]],
-    partition_keys: Optional[List[PartitionKey]],
-    input_deltas: pa.Table,
-    input_delta_type: DeltaType,
-    partition_values: Optional[List[Any]],
-    ds_mock_kwargs: Optional[Dict[str, Any]],
-    simulate_is_inplace: bool = False,
-) -> Tuple[Stream, Stream, Optional[Stream], str, str, str]:
-    
-    source_namespace, source_table_name, source_table_version = create_src_table_main(
-        sort_keys, partition_keys, input_deltas, ds_mock_kwargs
-    )
-
-    source_table_stream: Stream = metastore.get_stream(
-        namespace=source_namespace,
-        table_name=source_table_name,
-        table_version=source_table_version,
-        **ds_mock_kwargs,
-    )
-    
-    # Convert partition values to correct types
-    converted_partition_values = []
-    if partition_values and partition_keys:
-        for i, (value, pk) in enumerate(zip(partition_values, partition_keys)):
-            if pk.key_type == PartitionKeyType.INT:
-                converted_partition_values.append(int(value))
-            else:
-                converted_partition_values.append(value)
-    else:
-        converted_partition_values = partition_values
-    
-    staged_partition: Partition = metastore.stage_partition(
-        source_table_stream, converted_partition_values, **ds_mock_kwargs
-    )
-    metastore.commit_delta(
-        metastore.stage_delta(
-            input_deltas, staged_partition, input_delta_type, **ds_mock_kwargs
-        ),
-        **ds_mock_kwargs,
-    )
-    metastore.commit_partition(staged_partition, **ds_mock_kwargs)
-    source_table_stream_after_committed: Stream = metastore.get_stream(
-        namespace=source_namespace,
-        table_name=source_table_name,
-        table_version=source_table_version,
-        **ds_mock_kwargs,
-    )
-    
-    destination_table_namespace: Optional[str] = None
-    destination_table_name: Optional[str] = None
-    destination_table_version: Optional[str] = None
-    if not simulate_is_inplace:
-        (
-            destination_table_namespace,
-            destination_table_name,
-            destination_table_version,
-        ) = create_destination_table_main(sort_keys, partition_keys, input_deltas, ds_mock_kwargs)
-    else:
-        # not creating a table as in-place
-        destination_table_namespace = source_namespace
-        destination_table_name = source_table_name
-        destination_table_version = source_table_version
-
-    destination_table_stream: Stream = metastore.get_stream(
-        namespace=destination_table_namespace,
-        table_name=destination_table_name,
-        table_version=destination_table_version,
-        **ds_mock_kwargs,
-    )
-    return (
-        source_table_stream_after_committed,
-        destination_table_stream,
-        None,
-        source_namespace,
-        source_table_name,
-        source_table_version,
-    )
-
-
 @pytest.mark.parametrize(
     [
         "test_name",
@@ -448,7 +202,6 @@ def create_src_w_deltas_destination_plus_destination_main(
 def test_compact_partition_incremental_main(
     s3_resource: ServiceResource,
     main_deltacat_storage_kwargs: Dict[str, Any],
-    cleanup_database,
     test_name: str,
     primary_keys: Set[str],
     sort_keys: Dict[str, str],
